# back-end/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Dhumi Demo Production API
  version: 0.1.0
  description: |
    Accepted mock-first implementation contract for the Dhumi Marketplace Dataset
    and Scraper Library demo. Customers call only Dhumi. Bright Data credentials,
    dataset IDs, snapshot IDs and raw errors are never public fields.
    Customer operations use browser sessions. The 0070 refactor removes the
    three /v1/keys operations and secret-response recovery.

    This contract has no activation, customer plan, subscription, billing,
    credit or quota endpoints. Organization, verification and member workflows follow 0071.
    Live Bright Data traffic remains controlled by external evidence gates.
  contact:
    name: Dhumi Engineering
  license:
    name: Proprietary - internal implementation contract
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.dhumi.example
    description: Reserved documentation placeholder; deployment base URL is pending.
...
paths:
...
  /v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshSession
      x-dhumi-access: refresh_cookie_csrf
      summary: Rotate the browser session token family
      description: Refresh token is supplied through the approved secure cookie. Organization membership is not required.
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Session refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /v1/catalog/templates/{slug}:
    get:
      tags: [Catalogue]
      operationId: getTemplate
      x-dhumi-access: browser_user
      summary: Get one published Dhumi template
      description: Signed-in active users may read all-access templates without organization membership. Selected templates require an authorized organization with the access assignment.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/TemplateSlug'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTemplate'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/catalog/templates/{slug}/sample:
    get:
      tags: [Catalogue]
      operationId: getMarketplaceSample
      x-dhumi-access: browser_user
      summary: Read one page from a governed stored Marketplace sample
      description: Counts are relative to the immutable stored sample. Signed-in active users need no organization for all-access templates. Selected templates require authorized organization access. This operation makes no provider request.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/TemplateSlug'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '200':
          description: Stored sample page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplaceSampleQueryResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/catalog/templates/{slug}/sample/query:
    post:
      tags: [Catalogue]
      operationId: queryMarketplaceSample
      x-dhumi-access: browser_user
      summary: Filter and project a governed stored Marketplace sample
      description: Filtering is local to Dhumi's immutable stored sample and makes no provider request. Signed-in active users need no organization for all-access templates; selected templates require authorized organization access. Session CSRF remains required.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/TemplateSlug'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarketplaceSampleQueryInput' }
      responses:
        '200':
          description: Filtered and projected stored sample page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplaceSampleQueryResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/catalog/templates/{slug}/sample/downloads:
    post:
      tags: [Catalogue]
      operationId: authorizeMarketplaceSampleDownload
      x-dhumi-access: organization_member
      summary: Authorize a bounded download of a governed stored Marketplace sample
      description: >-
        Serializes the exact masked, filtered and selected projection of Dhumi's
        immutable stored sample as JSON or CSV. JSON preserves projected field
        values and types, but is not byte-identical source JSON. CSV is for
        spreadsheet viewing and prefixes formula-like cells with a tab inside
        quotes; it is not a lossless interchange format or a universal
        spreadsheet-security guarantee. This operation makes no provider
        request and does not create a Service, Run or outbox event.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/TemplateSlug'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarketplaceSampleDownloadInput' }
      responses:
        '201':
          description: Short-lived read-only sample-download authorization
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplaceSampleDownloadAuthorization' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/catalog/templates/{slug}/expert-enquiries:
    post:
      tags: [Catalogue]
      operationId: createMarketplaceExpertEnquiry
      x-dhumi-access: organization_member
      summary: Ask a Dhumi data expert about a Marketplace dataset
      description: >-
        Records one authenticated, Tenant-scoped enquiry against the exact
        Marketplace preview Template version. This operation creates no
        payment, entitlement, Service, Run, outbox event or provider request.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/TemplateSlug'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarketplaceExpertEnquiryInput' }
      responses:
        '201':
          description: Enquiry recorded, or the existing open enquiry returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplaceExpertEnquiry' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    BrowserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived Dhumi browser access token.
  headers:
...
    RetryAfter:
      description: Seconds before a safe retry may be attempted
      schema:
        type: integer
        minimum: 1
  parameters:
    OrganizationSelector:
      name: X-Dhumi-Organization
      in: header
      required: false
      description: Selector checked against active membership. Organization operations require selection with several memberships (400 ORGANIZATION_REQUIRED), and reject no membership (403 ORGANIZATION_MEMBERSHIP_REQUIRED). Browse needs no selector or membership for all-access templates. An inaccessible supplied selector gives 404 RESOURCE_NOT_FOUND.
      schema: { type: string, format: uuid }
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; Dhumi returns an accepted/generated ID.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque caller key scoped by authenticated actor/Tenant and operation.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Required for state-changing browser session operations.
      schema:
        type: string
        minLength: 16
        maxLength: 512
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    TemplateSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        minLength: 3
        maxLength: 100
...
  schemas:
...
    AuthSession:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_in, csrf_token]
      properties:
        access_token: { type: string, readOnly: true }
        token_type: { type: string, const: Bearer, readOnly: true }
        expires_in: { type: integer, minimum: 1, readOnly: true }
        csrf_token: { type: string, readOnly: true }
...
    ProductFamily:
      type: string
      enum: [marketplace_dataset, scraper_library]
    Availability:
      type: string
      enum: [available, preview_available, temporarily_unavailable, coming_soon]
    ServiceTemplate:
      type: object
      additionalProperties: false
      required: [slug, version, family, name, description, availability, presentation, configuration_schema, input_schema]
      properties:
        slug: { type: string, readOnly: true }
        version: { type: integer, minimum: 1, readOnly: true }
        family: { $ref: '#/components/schemas/ProductFamily' }
        name: { type: string, readOnly: true }
        description: { type: string, readOnly: true }
        availability: { $ref: '#/components/schemas/Availability' }
        presentation:
          $ref: '#/components/schemas/TemplatePresentation'
        configuration_schema:
          type: object
          readOnly: true
          description: Provider-neutral JSON Schema for settings saved in an immutable Service version. Changing product URLs do not belong here.
          additionalProperties: true
        input_schema:
          type: object
          readOnly: true
          description: Provider-neutral JSON Schema for input supplied separately to each Run, including Amazon product URLs.
          additionalProperties: true
        marketplace:
          $ref: '#/components/schemas/MarketplaceMetadata'
    TemplatePresentation:
      type: object
      additionalProperties: false
      required: [domain_slug, domain_name, category, icon_key, operation_group, operation_name, display_priority]
      properties:
        domain_slug:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
          readOnly: true
        domain_name: { type: string, minLength: 1, maxLength: 120, readOnly: true }
        category: { type: string, minLength: 1, maxLength: 80, readOnly: true }
        icon_key:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
          readOnly: true
          description: Key into Dhumi's approved frontend asset registry; never a provider-hosted URL.
        operation_group: { type: string, minLength: 1, maxLength: 120, readOnly: true }
        operation_name: { type: string, minLength: 1, maxLength: 120, readOnly: true }
        display_priority: { type: integer, minimum: 0, maximum: 1000000, readOnly: true }
...
    MarketplaceMetadata:
      type: object
      additionalProperties: false
      required: [record_count, record_count_as_of, sample, capabilities, fields, contact_modes]
      properties:
        record_count: { type: [integer, 'null'], minimum: 0, readOnly: true }
        record_count_as_of: { type: string, format: date-time, readOnly: true }
        sample:
          type: object
          additionalProperties: false
          required: [state, version, record_count, display_page_size, collected_at, expires_at, masking_notice]
          properties:
            state: { type: string, const: available, readOnly: true }
            version: { type: integer, minimum: 1, readOnly: true }
            record_count: { type: integer, minimum: 1, readOnly: true }
            display_page_size: { type: integer, minimum: 1, maximum: 100, readOnly: true }
            collected_at: { type: string, format: date-time, readOnly: true }
            expires_at: { type: string, format: date-time, readOnly: true }
            masking_notice: { type: string, minLength: 1, readOnly: true }
        capabilities:
          type: object
          additionalProperties: false
          required: [sample_query, sample_download, full_export]
          properties:
            sample_query: { $ref: '#/components/schemas/MarketplaceCapabilityState' }
            sample_download: { $ref: '#/components/schemas/MarketplaceCapabilityState' }
            full_export: { $ref: '#/components/schemas/MarketplaceCapabilityState' }
        fields:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/MarketplaceField' }
        contact_modes:
          type: array
          maxItems: 3
          items: { $ref: '#/components/schemas/MarketplaceContactMode' }
          readOnly: true
...
    MarketplaceFilterGroup:
      type: object
      additionalProperties: false
      required: [operator, filters]
      properties:
        operator: { type: string, enum: [and, or] }
        filters:
          type: array
          minItems: 1
          maxItems: 4
          items: { $ref: '#/components/schemas/MarketplaceFilterPredicate' }
...
    MarketplaceSampleQueryInput:
      type: object
      additionalProperties: false
      required: [expected_sample_version, selected_fields, page]
      properties:
        expected_sample_version: { type: integer, minimum: 1 }
        selected_fields:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 160 }
        filter:
          oneOf:
            - $ref: '#/components/schemas/MarketplaceFilterPredicate'
            - $ref: '#/components/schemas/MarketplaceFilterGroup'
        sort:
          type: array
          maxItems: 3
          items: { $ref: '#/components/schemas/MarketplaceSortItem' }
        page:
          type: object
          additionalProperties: false
          required: [limit]
          properties:
            limit: { type: integer, minimum: 1, maximum: 100, default: 30 }
            cursor: { type: [string, 'null'], minLength: 1, maxLength: 2048 }
    MarketplaceSampleQueryResult:
      type: object
      additionalProperties: false
      required: [template_slug, template_version, sample_version, sample_record_count, matches_in_sample, selected_fields, rows, masking_notice, page]
      properties:
        template_slug: { type: string, readOnly: true }
        template_version: { type: integer, minimum: 1, readOnly: true }
        sample_version: { type: integer, minimum: 1, readOnly: true }
        sample_record_count: { type: integer, minimum: 1, readOnly: true }
        matches_in_sample: { type: integer, minimum: 0, readOnly: true }
        selected_fields:
          type: array
          items: { type: string }
          readOnly: true
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
          readOnly: true
        masking_notice: { type: string, readOnly: true }
        page: { $ref: '#/components/schemas/Page' }
    MarketplaceSampleDownloadInput:
      type: object
      additionalProperties: false
      required: [expected_sample_version, selected_fields, format, record_limit]
      properties:
        expected_sample_version: { type: integer, minimum: 1 }
        selected_fields:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 160 }
        filter:
          oneOf:
            - $ref: '#/components/schemas/MarketplaceFilterPredicate'
            - $ref: '#/components/schemas/MarketplaceFilterGroup'
        sort:
          type: array
          maxItems: 3
          items: { $ref: '#/components/schemas/MarketplaceSortItem' }
        format:
          type: string
          enum: [json, csv]
          description: >-
            Use json for exact projected values and types. CSV is a
            spreadsheet-oriented representation that can alter formula-like
            cell values with a leading tab; use JSON for machine ingestion.
        record_limit:
          type: integer
          minimum: 1
          maximum: 100
          description: >-
            Dhumi's preview-download ceiling. It is intentionally independent
            of provider purchase/export limits and is further bounded by the
            current stored sample.
    MarketplaceSampleDownloadAuthorization:
      type: object
      additionalProperties: false
      required: [sample_version, format, record_count, content_type, byte_count, checksum, download_url, download_expires_at]
      properties:
        sample_version: { type: integer, minimum: 1, readOnly: true }
        format: { type: string, enum: [json, csv], readOnly: true }
        record_count: { type: integer, minimum: 0, readOnly: true }
        content_type: { type: string, readOnly: true }
        byte_count: { type: integer, minimum: 0, readOnly: true }
        checksum:
          type: string
          pattern: '^[a-f0-9]{64}$'
          readOnly: true
        download_url: { type: string, format: uri, readOnly: true }
        download_expires_at: { type: string, format: date-time, readOnly: true }
    MarketplaceExpertEnquiryInput:
      type: object
      additionalProperties: false
      required: [expected_template_version]
      properties:
        expected_template_version:
          type: integer
          minimum: 1
          description: Immutable Marketplace Template version shown to the customer.
    MarketplaceExpertEnquiry:
      type: object
      additionalProperties: false
      required: [id, template_slug, template_version, state, submitted_at]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        template_slug: { type: string, readOnly: true }
        template_version: { type: integer, minimum: 1, readOnly: true }
        state:
          type: string
          enum: [received, in_review, contacted, closed]
          readOnly: true
        submitted_at: { type: string, format: date-time, readOnly: true }
...
    Page:
      type: object
      additionalProperties: false
      required: [next_cursor, has_more]
      properties:
        next_cursor: { type: [string, 'null'], readOnly: true }
        has_more: { type: boolean, readOnly: true }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: [string, 'null'] }
        instance: { type: [string, 'null'], format: uri-reference }
        code: { $ref: '#/components/schemas/ProblemCode' }
        request_id: { type: string }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ProblemField' }
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PayloadTooLarge:
      description: Request body exceeds the accepted one-mebibyte limit
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnsupportedMediaType:
      description: Request body does not use a supported media type
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid Dhumi authentication
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not authorized or Tenant unavailable
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource does not exist or is not visible to this Tenant
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Idempotency or current-state conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Semantically invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Temporary Dhumi platform/product capacity limit; not a customer quota
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ServiceUnavailable:
      description: Dhumi or enabled product dependency temporarily unavailable
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }